feat : 검색창 지우기 함수 연결
